Move schema error tests from test_json_schema to test_types
